+++ vscode/package.json
@@ -797,12 +797,6 @@
           "markdownDescription": "Adds a Cody icon to the editor title menu for quick access to Cody commands.",
           "default": true
         },
-        "cody.experimental.chatPredictions": {
-          "order": 8,
-          "type": "boolean",
-          "default": false,
-          "markdownDescription": "Adds suggestions of possible relevant messages in the chat window."
-        },
         "cody.commandCodeLenses": {
           "order": 8,
           "type": "boolean",
