+++ vscode/src/chat/chat-view/SimpleChatPanelProvider.test.ts
@@ -3,10 +3,11 @@ import { URI } from 'vscode-uri'
 
 import type { Editor } from '@sourcegraph/cody-shared'
 
-import type { ContextItem } from './SimpleChatModel'
-import { contextFilesToContextItems } from './SimpleChatPanelProvider'
+import { type ContextItem } from './SimpleChatModel'
 
 import '../../testutils/vscode'
+
+import { contextFilesToContextItems } from './SimpleChatPanelProvider'
 
 describe('contextFilesToContextItems', () => {
     test('omits files that could not be read', async () => {
