+++ vscode/src/chat/chat-view/SimpleChatPanelProvider.ts
@@ -6,20 +6,13 @@ import {
     ConfigFeaturesSingleton,
     ContextWindowLimitError,
     hydrateAfterPostMessage,
-    isCodyIgnoredFile,
     isDefined,
     isDotCom,
     isError,
     isFileURI,
     isRateLimitError,
-    MAX_BYTES_PER_FILE,
-    NUM_CODE_RESULTS,
-    NUM_TEXT_RESULTS,
     reformatBotMessageForChat,
-    truncateTextNearestLine,
     Typewriter,
-    uriBasename,
-    type ActiveTextEditorSelectionRange,
     type ChatClient,
     type ChatMessage,
     type CodyCommand,
@@ -31,18 +24,13 @@ import {
     type Guardrails,
     type InteractionJSON,
     type Message,
-    type Result,
     type TranscriptJSON,
 } from '@sourcegraph/cody-shared'
 
 import type { View } from '../../../webviews/NavBar'
 import { newCodyCommandArgs, type CodyCommandArgs } from '../../commands'
 import type { CommandsController } from '../../commands/CommandsController'
-import {
-    createDisplayTextWithFileLinks,
-    createDisplayTextWithFileSelection,
-} from '../../commands/prompt/display-text'
-import { getContextForCommand } from '../../commands/utils/get-context'
+import { createDisplayTextWithFileLinks, createDisplayTextWithFileSelection } from '../../commands/prompt/display-text'
 import { getFullConfig } from '../../configuration'
 import { executeEdit } from '../../edit/execute'
 import {
@@ -52,10 +40,10 @@ import {
 } from '../../editor/utils/editor-context'
 import type { VSCodeEditor } from '../../editor/vscode-editor'
 import { ContextStatusAggregator } from '../../local-context/enhanced-context-status'
-import type { LocalEmbeddingsController } from '../../local-context/local-embeddings'
-import type { SymfRunner } from '../../local-context/symf'
-import { logDebug, logError } from '../../log'
-import type { AuthProvider } from '../../services/AuthProvider'
+import { type LocalEmbeddingsController } from '../../local-context/local-embeddings'
+import { type SymfRunner } from '../../local-context/symf'
+import { logDebug } from '../../log'
+import { type AuthProvider } from '../../services/AuthProvider'
 import { getProcessInfo } from '../../services/LocalAppDetector'
 import { localStorage } from '../../services/LocalStorageProvider'
 import { telemetryService } from '../../services/telemetry'
@@ -69,32 +57,20 @@ import {
 } from '../../services/utils/codeblock-action-tracker'
 import { openExternalLinks, openLocalFileWithRange } from '../../services/utils/workspace-action'
 import { TestSupport } from '../../test-support'
-import type { CachedRemoteEmbeddingsClient } from '../CachedRemoteEmbeddingsClient'
-import type { MessageErrorType } from '../MessageProvider'
-import type {
-    AuthStatus,
-    ConfigurationSubsetForWebview,
-    ExtensionMessage,
-    LocalEnv,
-    WebviewMessage,
-} from '../protocol'
 import { countGeneratedCode } from '../utils'
 
-import { getChatPanelTitle, openFile, stripContextWrapper } from './chat-helpers'
+import { getChatPanelTitle, openFile, stripContextWrapper, viewRangeToRange } from './chat-helpers'
 import { ChatHistoryManager } from './ChatHistoryManager'
 import { addWebviewViewHTML, CodyChatPanelViewType } from './ChatManager'
 import type { ChatViewProviderWebview, Config } from './ChatPanelsManager'
 import { CodebaseStatusProvider } from './CodebaseStatusProvider'
+import { getCommandContext, getEnhancedContext } from './context'
 import { InitDoer } from './InitDoer'
-import { DefaultPrompter, type IContextProvider, type IPrompter } from './prompt'
-import {
-    SimpleChatModel,
-    toViewMessage,
-    type ContextItem,
-    type MessageWithContext,
-} from './SimpleChatModel'
-
-const isAgentTesting = process.env.CODY_SHIM_TESTING === 'true'
+import { ContextItem, MessageWithContext, SimpleChatModel, toViewMessage } from './SimpleChatModel'
+import { CachedRemoteEmbeddingsClient } from '../CachedRemoteEmbeddingsClient'
+import { AuthStatus, ChatSubmitType, ConfigurationSubsetForWebview, ExtensionMessage, LocalEnv, WebviewMessage } from '../protocol'
+import { CommandPrompter, DefaultPrompter, IPrompter } from './prompt'
+import { MessageErrorType } from '../MessageProvider'
 
 interface SimpleChatPanelProviderOptions {
     config: Config
@@ -118,20 +94,7 @@ export interface ChatSession {
 }
 
 export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
-    private _webviewPanel?: vscode.WebviewPanel
-    public get webviewPanel(): vscode.WebviewPanel | undefined {
-        return this._webviewPanel
-    }
-    private _webview?: ChatViewProviderWebview
-    public get webview(): ChatViewProviderWebview | undefined {
-        return this._webview
-    }
-    private initDoer = new InitDoer<boolean | undefined>()
-
     private chatModel: SimpleChatModel
-
-    private extensionUri: vscode.Uri
-    private disposables: vscode.Disposable[] = []
 
     private config: Config
     private readonly authProvider: AuthProvider
@@ -147,9 +110,13 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
     private readonly commandsController?: CommandsController
 
     private history = new ChatHistoryManager()
-    private prompter: IPrompter = new DefaultPrompter()
-
     private contextFilesQueryCancellation?: vscode.CancellationTokenSource
+
+    // A unique identifier for this SimpleChatPanelProvider instance used to identify it to
+    // container classes that need a handle to this specific panel provider.
+    public get sessionID(): string {
+        return this.chatModel.sessionID
+    }
 
     constructor({
         config,
@@ -175,7 +142,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         this.commandsController = commandsController
         this.editor = editor
         this.treeView = treeView
-        this.chatModel = new SimpleChatModel(this.selectModel(models))
+        this.chatModel = new SimpleChatModel(selectModel(authProvider, models))
         this.guardrails = guardrails
 
         commandsController?.setEnableExperimentalCommands(config.internalUnstable)
@@ -203,43 +170,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         this.disposables.push(this.contextStatusAggregator.addProvider(this.codebaseStatusProvider))
     }
 
-    public get sessionID(): string {
-        return this.chatModel.sessionID
-    }
-
-    // Select the chat model to use in Chat
-    private selectModel(models: ChatModelProvider[]): string {
-        const authStatus = this.authProvider.getAuthStatus()
-        // Free user can only use the default model
-        if (authStatus.isDotCom && authStatus.userCanUpgrade) {
-            return models[0].model
-        }
-        // Check for the last selected model
-        const lastSelectedModelID = localStorage.get('model')
-        if (lastSelectedModelID) {
-            // If the last selected model exists in the list of models then we return it
-            const model = models.find(m => m.model === lastSelectedModelID)
-            if (model) {
-                return lastSelectedModelID
-            }
-        }
-        // If the user has not selected a model before then we return the default model
-        const defaultModel = models.find(m => m.default) || models[0]
-        if (!defaultModel) {
-            throw new Error('No chat model found in server-provided config')
-        }
-        return defaultModel.model
-    }
-
-    private completionCanceller?: () => void
-
-    private cancelInProgressCompletion(): void {
-        if (this.completionCanceller) {
-            this.completionCanceller()
-            this.completionCanceller = undefined
-        }
-    }
-
     /**
      * Creates the webview panel for the Cody chat interface if it doesn't already exist.
      */
@@ -256,7 +186,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
 
         const viewType = CodyChatPanelViewType
         const panelTitle =
-            this.history.getChat(this.authProvider.getAuthStatus(), this.sessionID)?.chatTitle ||
+            this.history.getChat(this.authProvider.getAuthStatus(), this.chatModel.sessionID)?.chatTitle ||
             getChatPanelTitle(lastQuestion)
         const viewColumn = activePanelViewColumn || vscode.ViewColumn.Beside
         const webviewPath = vscode.Uri.joinPath(this.extensionUri, 'dist', 'webviews')
@@ -339,6 +269,18 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         return panel
     }
 
+    public async setWebviewView(view: View): Promise<void> {
+        if (!this.webviewPanel) {
+            await this.createWebviewPanel()
+        }
+        this.webviewPanel?.reveal()
+
+        await this.postMessage({
+            type: 'view',
+            messages: view,
+        })
+    }
+
     private postContextStatusToWebView(): void {
         logDebug(
             'SimpleChatPanelProvider',
@@ -350,18 +292,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             context: {
                 groups: this.contextStatusAggregator.status,
             },
-        })
-    }
-
-    public async setWebviewView(view: View): Promise<void> {
-        if (!this.webviewPanel) {
-            await this.createWebviewPanel()
-        }
-        this.webviewPanel?.reveal()
-
-        await this.postMessage({
-            type: 'view',
-            messages: view,
         })
     }
 
@@ -382,12 +312,12 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                 break
             case 'submit': {
                 const requestID = uuid.v4()
-                await this.handleHumanMessageSubmitted(
+                await this.handleNewUserMessage(
                     requestID,
                     message.text,
                     message.submitType,
                     message.contextFiles ?? [],
-                    message.addEnhancedContext || false
+                    message.addEnhancedContext ?? false
                 )
                 break
             }
@@ -494,6 +424,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         }
     }
 
+    private initDoer = new InitDoer<boolean | undefined>()
     private async onInitialized(): Promise<void> {
         // HACK: this call is necessary to get the webview to set the chatID state,
         // which is necessary on deserialization. It should be invoked before the
@@ -503,7 +434,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             type: 'transcript',
             messages: [],
             isMessageInProgress: false,
-            chatID: this.sessionID,
+            chatID: this.chatModel.sessionID,
         })
 
         this.postChatModels()
@@ -512,6 +443,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         this.initDoer.signalInitialized()
     }
 
+    private disposables: vscode.Disposable[] = []
     public dispose(): void {
         vscode.Disposable.from(...this.disposables).dispose()
         this.disposables = []
@@ -597,58 +529,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         })
     }
 
-    /**
-     * Handles a message submitted by the user.
-     *
-     * Validates the message, checks for slash commands, edit commands,
-     * and sends the message to be handled like a regular chat request.
-     */
-    public async handleHumanMessageSubmitted(
-        requestID: string,
-        text: string,
-        submitType: 'user' | 'user-newchat' | 'suggestion' | 'example',
-        userContextFiles: ContextFile[],
-        addEnhancedContext: boolean
-    ): Promise<void> {
-        if (submitType === 'suggestion') {
-            const args = { requestID }
-            telemetryService.log('CodyVSCodeExtension:chatPredictions:used', args, { hasV2Event: true })
-        }
-        // If this is a slash command, run it with custom command instead
-        if (text.startsWith('/')) {
-            if (text.match(/^\/r(eset)?$/)) {
-                return this.clearAndRestartSession()
-            }
-            if (text.match(/^\/edit(\s)?/)) {
-                return executeEdit({ instruction: text.replace(/^\/(edit)/, '').trim() }, 'chat')
-            }
-            if (text === '/commands-settings') {
-                // User has clicked the settings button for commands
-                return vscode.commands.executeCommand('cody.settings.commands')
-            }
-            const commandArgs = newCodyCommandArgs({ source: 'chat', requestID })
-            const command = await this.commandsController?.startCommand(text, commandArgs)
-            if (command) {
-                return this.handleCommands(command, commandArgs)
-            }
-        }
-
-        // HACK
-        if (submitType === 'user-newchat' && !this.chatModel.isEmpty()) {
-            await this.clearAndRestartSession()
-        }
-
-        await this.handleChatRequest(requestID, text, submitType, userContextFiles, addEnhancedContext)
-    }
-
-    /**
-     * Handles executing a chat command from the user.
-     *
-     * Validates the command, checks for edit commands,
-     * generates a chat request from the command,
-     * and sends it to be handled like a regular chat request.
-     */
-    public async handleCommands(command: CodyCommand, args: CodyCommandArgs): Promise<void> {
+    public async handleCommand(command: CodyCommand, args: CodyCommandArgs): Promise<void> {
         // If it's not a ask command, it's a fixup command. If it's a fixup request, we can exit early
         // This is because startCommand will start the CommandRunner,
         // which would send all fixup requests to the FixupController
@@ -656,111 +537,228 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             return
         }
 
-        // If editor is not active, post error and return early
-        if (command && !this.editor.getActiveTextEditorSelectionOrVisibleContent()) {
-            if (
-                command.context?.selection ||
-                command.context?.currentFile ||
-                command.context?.currentDir
-            ) {
-                return this.postError(
-                    new Error('Command failed. Please open a file and try again.'),
-                    'transcript'
-                )
+        if (!this.editor.getActiveTextEditorSelectionOrVisibleContent()) {
+            if (command.context?.selection || command.context?.currentFile || command.context?.currentDir) {
+                return this.postError(new Error('Command failed. Please open a file and try again.'), 'transcript')
             }
         }
 
         const inputText = [command.slashCommand, command.additionalInput].join(' ')?.trim()
+        const displayText = createDisplayTextWithFileSelection(
+            inputText,
+            this.editor.getActiveTextEditorSelectionOrEntireFile()
+        )
+        const promptText = command.prompt
+        this.chatModel.addHumanMessage({ text: promptText }, displayText)
+        this.updateWebviewPanelTitle(inputText)
+        await this.saveSession(inputText)
 
-        await this.handleChatRequest(args.requestID, inputText, 'user', [], false, command)
+        // trigger the context progress indicator
+        this.postViewTranscript({ speaker: 'assistant' })
+
+        const prompt = await this.buildPrompt(new CommandPrompter(() => getCommandContext(this.editor, command)))
+        this.streamAssistantResponse(args.requestID, prompt)
+    }
+
+    public async handleNewUserMessage(
+        requestID: string,
+        inputText: string,
+        submitType: ChatSubmitType,
+        userContextFiles: ContextFile[],
+        addEnhancedContext: boolean
+    ): Promise<void> {
+        // DEPRECATED (remove after slash commands are removed)
+        // If this is a slash command, run it with custom command instead
+        if (inputText.startsWith('/')) {
+            if (inputText.match(/^\/r(eset)?$/)) {
+                return this.clearAndRestartSession()
+            }
+            if (inputText.match(/^\/edit(\s)?/)) {
+                return executeEdit({ instruction: inputText.replace(/^\/(edit)/, '').trim() }, 'chat')
+            }
+            if (inputText === '/commands-settings') {
+                // User has clicked the settings button for commands
+                return vscode.commands.executeCommand('cody.settings.commands')
+            }
+            const commandArgs = newCodyCommandArgs({ source: 'chat', requestID })
+            const command = await this.commandsController?.startCommand(inputText, commandArgs)
+            if (command) {
+                return this.handleCommand(command, commandArgs)
+            }
+        }
+
+        if (submitType === 'user-newchat' && !this.chatModel.isEmpty()) {
+            await this.clearAndRestartSession()
+        }
+
+        const displayText = userContextFiles?.length
+            ? createDisplayTextWithFileLinks(inputText, userContextFiles)
+            : inputText
+        const promptText = inputText
+        this.chatModel.addHumanMessage({ text: promptText }, displayText)
+        this.updateWebviewPanelTitle(inputText)
+        await this.saveSession(inputText)
+
+        // trigger the context progress indicator
+        this.postViewTranscript({ speaker: 'assistant' })
+
+        const userContextItems = await contextFilesToContextItems(this.editor, userContextFiles || [], true)
+        const prompter = new DefaultPrompter(
+            userContextItems,
+            addEnhancedContext
+                ? query =>
+                      getEnhancedContext(
+                          this.config.useContext,
+                          this.editor,
+                          this.embeddingsClient,
+                          this.localEmbeddings,
+                          this.config.experimentalSymfContext ? this.symf : null,
+                          this.codebaseStatusProvider,
+                          query
+                      )
+                : undefined
+        )
+        const sendTelemetry = (contextSummary: {}): void => {
+            const authStatus = this.authProvider.getAuthStatus()
+            const properties = {
+                requestID,
+                chatModel: this.chatModel.modelID,
+                contextSummary,
+            }
+
+            // Only log chat-question event if it is not a command to avoid double logging for commands
+            telemetryService.log('CodyVSCodeExtension:chat-question:executed', properties, {
+                hasV2Event: true,
+            })
+            telemetryRecorder.recordEvent('cody.chat-question', 'executed', {
+                metadata: {
+                    ...contextSummary,
+                    // Flag indicating this is a transcript event to go through ML data pipeline. Only for DotCom users
+                    // See https://github.com/sourcegraph/sourcegraph/pull/59524
+                    recordsPrivateMetadataTranscript: authStatus.endpoint && isDotCom(authStatus.endpoint) ? 1 : 0,
+                },
+                privateMetadata: {
+                    properties,
+                    // 🚨 SECURITY: chat transcripts are to be included only for DotCom users AND for V2 telemetry
+                    // V2 telemetry exports privateMetadata only for DotCom users
+                    // the condition below is an aditional safegaurd measure
+                    promptText: authStatus.endpoint && isDotCom(authStatus.endpoint) ? promptText : undefined,
+                },
+            })
+        }
+
+        try {
+            const prompt = await this.buildPrompt(prompter, sendTelemetry)
+            this.streamAssistantResponse(requestID, prompt)
+        } catch (error) {
+            if (isRateLimitError(error)) {
+                this.postError(error, 'transcript')
+            } else {
+                this.postError(isError(error) ? error : new Error(`Error generating assistant response: ${error}`))
+            }
+        }
     }
 
     /**
-     * Handles a chat request from chat input or a command.
-     *
-     * Saves the chat session, posts a transcript update, generates the
-     * assistant's response, logs telemetry, and updates the panel title.
+     * Constructs the prompt and updates the UI with the context used in the prompt.
      */
-    private async handleChatRequest(
-        requestID: string,
-        inputText: string,
-        submitType: 'user' | 'suggestion' | 'example' | 'user-newchat',
-        userContextFiles: ContextFile[],
-        addEnhancedContext: boolean,
-        command?: CodyCommand
-    ): Promise<void> {
-        // Display text is the text we will display to the user in the Chat UI
-        // - Append @-files to the display text if we have any
-        // - Append @-file selection for commands
-        // Otherwise, use the input text
-        const displayText = userContextFiles?.length
-            ? createDisplayTextWithFileLinks(inputText, userContextFiles)
-            : command
-              ? createDisplayTextWithFileSelection(
-                      inputText,
-                      this.editor.getActiveTextEditorSelectionOrEntireFile()
-                  )
-              : inputText
-        // The text we will use to send to LLM
-        const promptText = command ? command.prompt : inputText
-        this.chatModel.addHumanMessage({ text: promptText }, displayText)
+    private async buildPrompt(prompter: IPrompter, sendTelemetry?: (contextSummary: {}) => void): Promise<Message[]> {
+        const { prompt, contextLimitWarnings, newContextUsed } = await prompter.makePrompt(
+            this.chatModel,
+            getContextWindowForModel(this.authProvider.getAuthStatus(), this.chatModel.modelID)
+        )
 
-        await this.saveSession(inputText)
-        // trigger the context progress indicator
+        // Update UI based on prompt construction
+        this.chatModel.setNewContextUsed(newContextUsed)
+        if (contextLimitWarnings.length > 0) {
+            const warningMsg = contextLimitWarnings
+                .map(w => (w.trim().endsWith('.') ? w.trim() : w.trim() + '.'))
+                .join(' ')
+            this.postError(new ContextWindowLimitError(warningMsg), 'transcript')
+        }
+
+        if (sendTelemetry) {
+            // Create a summary of how many code snippets of each context source are being
+            // included in the prompt
+            const contextSummary: { [key: string]: number } = {}
+            for (const { source } of newContextUsed) {
+                if (!source) {
+                    continue
+                }
+                if (contextSummary[source]) {
+                    contextSummary[source] += 1
+                } else {
+                    contextSummary[source] = 1
+                }
+            }
+            sendTelemetry(contextSummary)
+        }
+
+        return prompt
+    }
+
+    private streamAssistantResponse(requestID: string, prompt: Message[]): void {
         this.postViewTranscript({ speaker: 'assistant' })
-        await this.generateAssistantResponse(
-            requestID,
-            userContextFiles,
-            addEnhancedContext,
-            contextSummary => {
-                if (submitType !== 'user' && submitType !== 'user-newchat') {
-                    return
-                }
 
-                const authStatus = this.authProvider.getAuthStatus()
-
-                const properties = {
-                    requestID,
-                    chatModel: this.chatModel.modelID,
-                    contextSummary,
-                }
-
-                // Only log chat-question event if it is not a command to avoid double logging for commands
-                if (!command) {
-                    telemetryService.log('CodyVSCodeExtension:chat-question:executed', properties, {
-                        hasV2Event: true,
-                    })
-                    telemetryRecorder.recordEvent('cody.chat-question', 'executed', {
-                        metadata: {
-                            ...contextSummary,
-                            // Flag indicating this is a transcript event to go through ML data pipeline. Only for DotCom users
-                            // See https://github.com/sourcegraph/sourcegraph/pull/59524
-                            recordsPrivateMetadataTranscript:
-                                authStatus.endpoint && isDotCom(authStatus.endpoint) ? 1 : 0,
-                        },
-                        privateMetadata: {
-                            properties,
-                            // 🚨 SECURITY: chat transcripts are to be included only for DotCom users AND for V2 telemetry
-                            // V2 telemetry exports privateMetadata only for DotCom users
-                            // the condition below is an aditional safegaurd measure
-                            promptText:
-                                authStatus.endpoint && isDotCom(authStatus.endpoint)
-                                    ? promptText
-                                    : undefined,
+        this.sendLLMRequest(prompt, {
+            update: content => {
+                this.postViewTranscript(
+                    toViewMessage({
+                        message: {
+                            speaker: 'assistant',
+                            text: content,
                         },
                     })
+                )
+            },
+            close: content => {
+                this.addBotMessage(requestID, content)
+            },
+            error: (partialResponse, error) => {
+                if (!isAbortError(error)) {
+                    this.postError(error, 'transcript')
+                }
+                try {
+                    // We should still add the partial response if there was an error
+                    // This'd throw an error if one has already been added
+                    this.addBotMessage(requestID, partialResponse)
+                } catch {
+                    console.error('Streaming Error', error)
                 }
             },
-            command
-        )
-        // Set the title of the webview panel
-        this.updateWebviewPanelTitle(inputText)
+        })
     }
 
     private async handleEdit(requestID: string, text: string): Promise<void> {
         this.chatModel.updateLastHumanMessage({ text })
         this.postViewTranscript()
-        await this.generateAssistantResponse(requestID)
+
+        const prompter = new DefaultPrompter(
+            [], // TODO(beyang): support user context items in the edit input
+            (
+                query // TODO(beyang): get useEnhancedContext
+            ) =>
+                getEnhancedContext(
+                    this.config.useContext,
+                    this.editor,
+                    this.embeddingsClient,
+                    this.localEmbeddings,
+                    this.config.experimentalSymfContext ? this.symf : null,
+                    this.codebaseStatusProvider,
+                    query
+                )
+        )
+
+        try {
+            const prompt = await this.buildPrompt(prompter)
+            this.streamAssistantResponse(requestID, prompt)
+        } catch (error) {
+            if (isRateLimitError(error)) {
+                this.postError(error, 'transcript')
+            } else {
+                this.postError(isError(error) ? error : new Error(`Error generating assistant response: ${error}`))
+            }
+        }
     }
 
     private async postViewConfig(): Promise<void> {
@@ -782,107 +780,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             workspaceFolderUris,
         })
         logDebug('SimpleChatPanelProvider', 'updateViewConfig', { verbose: configForWebview })
-    }
-
-    private async generateAssistantResponse(
-        requestID: string,
-        userContextFiles?: ContextFile[],
-        addEnhancedContext = true,
-        sendTelemetry?: (contextSummary: Record<string, number>) => void,
-        command?: CodyCommand
-    ): Promise<void> {
-        try {
-            const contextWindowBytes = getContextWindowForModel(
-                this.authProvider.getAuthStatus(),
-                this.chatModel.modelID
-            )
-
-            const userContextItems = await contextFilesToContextItems(
-                this.editor,
-                userContextFiles || [],
-                true
-            )
-            const contextProvider = new ContextProvider(
-                userContextItems,
-                this.editor,
-                this.embeddingsClient,
-                this.localEmbeddings,
-                this.config.experimentalSymfContext ? this.symf : null,
-                this.codebaseStatusProvider
-            )
-            const { prompt, contextLimitWarnings, newContextUsed } = await this.prompter.makePrompt(
-                this.chatModel,
-                contextProvider,
-                addEnhancedContext,
-                contextWindowBytes,
-                command
-            )
-
-            this.chatModel.setNewContextUsed(newContextUsed)
-
-            if (contextLimitWarnings.length > 0) {
-                const warningMsg = contextLimitWarnings
-                    .map(w => (w.trim().endsWith('.') ? w.trim() : `${w.trim()}.`))
-                    .join(' ')
-                this.postError(new ContextWindowLimitError(warningMsg), 'transcript')
-            }
-
-            if (sendTelemetry) {
-                // Create a summary of how many code snippets of each context source are being
-                // included in the prompt
-                const contextSummary: { [key: string]: number } = {}
-                for (const { source } of newContextUsed) {
-                    if (!source) {
-                        continue
-                    }
-                    if (contextSummary[source]) {
-                        contextSummary[source] += 1
-                    } else {
-                        contextSummary[source] = 1
-                    }
-                }
-                sendTelemetry(contextSummary)
-            }
-
-            this.postViewTranscript({ speaker: 'assistant' })
-
-            await this.sendLLMRequest(prompt, {
-                update: content => {
-                    this.postViewTranscript(
-                        toViewMessage({
-                            message: {
-                                speaker: 'assistant',
-                                text: content,
-                            },
-                            newContextUsed,
-                        })
-                    )
-                },
-                close: content => {
-                    this.addBotMessage(requestID, content)
-                },
-                error: (partialResponse, error) => {
-                    if (!isAbortError(error)) {
-                        this.postError(error, 'transcript')
-                    }
-                    try {
-                        // We should still add the partial response if there was an error
-                        // This'd throw an error if one has already been added
-                        this.addBotMessage(requestID, partialResponse)
-                    } catch {
-                        console.error('Streaming Error', error)
-                    }
-                },
-            })
-        } catch (error) {
-            if (isRateLimitError(error)) {
-                this.postError(error, 'transcript')
-            } else {
-                this.postError(
-                    isError(error) ? error : new Error(`Error generating assistant response: ${error}`)
-                )
-            }
-        }
     }
 
     /**
@@ -938,6 +835,13 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                     typewriter.stop(message.error)
                 }
             }
+        }
+    }
+    private completionCanceller?: () => void
+    private cancelInProgressCompletion(): void {
+        if (this.completionCanceller) {
+            this.completionCanceller()
+            this.completionCanceller = undefined
         }
     }
 
@@ -1007,13 +911,10 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             type: 'transcript',
             messages,
             isMessageInProgress: !!messageInProgress,
-            chatID: this.sessionID,
+            chatID: this.chatModel.sessionID,
         })
 
-        const chatTitle = this.history.getChat(
-            this.authProvider.getAuthStatus(),
-            this.sessionID
-        )?.chatTitle
+        const chatTitle = this.history.getChat(this.authProvider.getAuthStatus(), this.chatModel.sessionID)?.chatTitle
         if (chatTitle) {
             this.handleChatTitle(chatTitle)
             return
@@ -1131,6 +1032,19 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
     }
 
     /**
+     * Webview state
+     */
+    private extensionUri: vscode.Uri
+    private _webviewPanel?: vscode.WebviewPanel
+    public get webviewPanel(): vscode.WebviewPanel | undefined {
+        return this._webviewPanel
+    }
+    private _webview?: ChatViewProviderWebview
+    public get webview(): ChatViewProviderWebview | undefined {
+        return this._webview
+    }
+
+    /**
      * Posts a message to the webview, pending initialization.
      *
      * cody-invariant: this.webview?.postMessage should never be invoked directly
@@ -1143,7 +1057,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
     private updateWebviewPanelTitle(title: string): void {
         if (this.webviewPanel) {
             this.webviewPanel.title =
-                this.history.getChat(this.authProvider.getAuthStatus(), this.sessionID)?.chatTitle ||
+                this.history.getChat(this.authProvider.getAuthStatus(), this.chatModel.sessionID)?.chatTitle ||
                 getChatPanelTitle(title)
         }
     }
@@ -1160,458 +1074,32 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
     }
 }
 
-class ContextProvider implements IContextProvider {
-    constructor(
-        private userContext: ContextItem[],
-        private editor: VSCodeEditor,
-        private embeddingsClient: CachedRemoteEmbeddingsClient | null,
-        private localEmbeddings: LocalEmbeddingsController | null,
-        private symf: SymfRunner | null,
-        private codebaseStatusProvider: CodebaseStatusProvider
-    ) {}
-
-    public getExplicitContext(): ContextItem[] {
-        return this.userContext
-    }
-
-    private getUserAttentionContext(): ContextItem[] {
-        return this.getVisibleEditorContext()
-    }
-
-    public async getSmartSelectionContext(): Promise<ContextItem[]> {
-        const smartSelection = await this.editor.getActiveTextEditorSmartSelection()
-        const selection = smartSelection || this.editor.getActiveTextEditorSelectionOrVisibleContent()
-        if (!selection?.selectedText || isCodyIgnoredFile(selection.fileUri)) {
-            return []
+async function newChatModelfromTranscriptJSON(json: TranscriptJSON, modelID: string): Promise<SimpleChatModel> {
+    const messages: MessageWithContext[][] = json.interactions.map(
+        (interaction: InteractionJSON): MessageWithContext[] => {
+            return [
+                {
+                    message: {
+                        speaker: 'human',
+                        text: interaction.humanMessage.text,
+                    },
+                    displayText: interaction.humanMessage.displayText,
+                    newContextUsed: deserializedContextFilesToContextItems(
+                        interaction.usedContextFiles,
+                        interaction.fullContext
+                    ),
+                },
+                {
+                    message: {
+                        speaker: 'assistant',
+                        text: interaction.assistantMessage.text,
+                    },
+                    displayText: interaction.assistantMessage.displayText,
+                },
+            ]
         }
-        let range: vscode.Range | undefined
-        if (selection.selectionRange) {
-            range = new vscode.Range(
-                selection.selectionRange.start.line,
-                selection.selectionRange.start.character,
-                selection.selectionRange.end.line,
-                selection.selectionRange.end.character
-            )
-        }
-
-        return [
-            {
-                text: selection.selectedText,
-                uri: selection.fileUri,
-                range,
-                source: 'selection',
-            },
-        ]
-    }
-
-    public getCurrentSelectionContext(): ContextItem[] {
-        const selection = this.editor.getActiveTextEditorSelection()
-        if (!selection?.selectedText || isCodyIgnoredFile(selection.fileUri)) {
-            return []
-        }
-        let range: vscode.Range | undefined
-        if (selection.selectionRange) {
-            range = new vscode.Range(
-                selection.selectionRange.start.line,
-                selection.selectionRange.start.character,
-                selection.selectionRange.end.line,
-                selection.selectionRange.end.character
-            )
-        }
-
-        return [
-            {
-                text: selection.selectedText,
-                uri: selection.fileUri,
-                range,
-                source: 'selection',
-            },
-        ]
-    }
-
-    private getVisibleEditorContext(): ContextItem[] {
-        const visible = this.editor.getActiveTextEditorVisibleContent()
-        const fileUri = visible?.fileUri
-        if (!visible || !fileUri) {
-            return []
-        }
-        if (isCodyIgnoredFile(fileUri) || !visible.content.trim()) {
-            return []
-        }
-        return [
-            {
-                text: visible.content,
-                uri: fileUri,
-                source: 'editor',
-            },
-        ]
-    }
-
-    public async getEnhancedContext(text: string): Promise<ContextItem[]> {
-        const config = vscode.workspace.getConfiguration('cody')
-        const useContextConfig = config.get('useContext')
-
-        const searchContext: ContextItem[] = []
-
-        // use user attention context only if config is set to none
-        if (useContextConfig === 'none') {
-            logDebug('SimpleChatPanelProvider', 'getEnhancedContext > none')
-            searchContext.push(...this.getUserAttentionContext())
-            return searchContext
-        }
-
-        let hasEmbeddingsContext = false
-        // Get embeddings context if useContext Config is not set to 'keyword' only
-        if (useContextConfig !== 'keyword') {
-            logDebug('SimpleChatPanelProvider', 'getEnhancedContext > embeddings (start)')
-            const localEmbeddingsResults = this.searchEmbeddingsLocal(text)
-            const remoteEmbeddingsResults = this.searchEmbeddingsRemote(text)
-            try {
-                const r = await localEmbeddingsResults
-                hasEmbeddingsContext = hasEmbeddingsContext || r.length > 0
-                searchContext.push(...r)
-            } catch (error) {
-                logDebug('SimpleChatPanelProvider', 'getEnhancedContext > local embeddings', error)
-            }
-            try {
-                const r = await remoteEmbeddingsResults
-                hasEmbeddingsContext = hasEmbeddingsContext || r.length > 0
-                searchContext.push(...r)
-            } catch (error) {
-                logDebug('SimpleChatPanelProvider', 'getEnhancedContext > remote embeddings', error)
-            }
-            logDebug('SimpleChatPanelProvider', 'getEnhancedContext > embeddings (end)')
-        }
-
-        // Fallback to symf if embeddings provided no results or if useContext is set to 'keyword' specifically
-        if (!hasEmbeddingsContext && this.symf) {
-            logDebug('SimpleChatPanelProvider', 'getEnhancedContext > search')
-            try {
-                searchContext.push(...(await this.searchSymf(text)))
-            } catch (error) {
-                // TODO(beyang): handle this error better
-                logDebug('SimpleChatPanelProvider.getEnhancedContext', 'searchSymf error', error)
-            }
-        }
-
-        const priorityContext: ContextItem[] = []
-        const selectionContext = this.getCurrentSelectionContext()
-        if (selectionContext.length > 0) {
-            priorityContext.push(...selectionContext)
-        } else if (this.needsUserAttentionContext(text)) {
-            // Query refers to current editor
-            priorityContext.push(...this.getUserAttentionContext())
-        } else if (this.needsReadmeContext(text)) {
-            // Query refers to project, so include the README
-            let containsREADME = false
-            for (const contextItem of searchContext) {
-                const basename = uriBasename(contextItem.uri)
-                if (
-                    basename.toLocaleLowerCase() === 'readme' ||
-                    basename.toLocaleLowerCase().startsWith('readme.')
-                ) {
-                    containsREADME = true
-                    break
-                }
-            }
-            if (!containsREADME) {
-                priorityContext.push(...(await this.getReadmeContext()))
-            }
-        }
-
-        return priorityContext.concat(searchContext)
-    }
-
-    public async getCommandContext(command: CodyCommand): Promise<ContextItem[]> {
-        logDebug('SimpleChatPanelProvider.getCommandContext', command.slashCommand)
-
-        const contextMessages: ContextMessage[] = []
-        const contextItems: ContextItem[] = []
-
-        contextMessages.push(...(await getContextForCommand(this.editor, command)))
-        // Turn ContextMessages to ContextItems
-        for (const msg of contextMessages) {
-            if (msg.file?.uri && msg.file?.content) {
-                contextItems.push({
-                    uri: msg.file?.uri,
-                    text: msg.file?.content,
-                    range: viewRangeToRange(msg.file?.range),
-                    source: msg.file?.source || 'editor',
-                })
-            }
-        }
-        // Add codebase ContextItems last
-        if (command.context?.codebase) {
-            contextItems.push(...(await this.getEnhancedContext(command.prompt)))
-        }
-
-        return contextItems
-    }
-
-    /**
-     * Uses symf to conduct a local search within the current workspace folder
-     */
-    private async searchSymf(userText: string, blockOnIndex = false): Promise<ContextItem[]> {
-        if (!this.symf) {
-            return []
-        }
-        const workspaceRoot = this.editor.getWorkspaceRootUri()
-        if (!workspaceRoot || !isFileURI(workspaceRoot)) {
-            return []
-        }
-
-        const indexExists = await this.symf.getIndexStatus(workspaceRoot)
-        if (indexExists !== 'ready' && !blockOnIndex) {
-            void this.symf.ensureIndex(workspaceRoot, { hard: false })
-            return []
-        }
-
-        const r0 = (await this.symf.getResults(userText, [workspaceRoot])).flatMap(async results => {
-            const items = (await results).flatMap(
-                async (result: Result): Promise<ContextItem[] | ContextItem> => {
-                    if (isCodyIgnoredFile(result.file)) {
-                        return []
-                    }
-                    const range = new vscode.Range(
-                        result.range.startPoint.row,
-                        result.range.startPoint.col,
-                        result.range.endPoint.row,
-                        result.range.endPoint.col
-                    )
-
-                    let text: string | undefined
-                    try {
-                        text = await this.editor.getTextEditorContentForFile(result.file, range)
-                        if (!text) {
-                            return []
-                        }
-                    } catch (error) {
-                        logError(
-                            'SimpleChatPanelProvider.searchSymf',
-                            `Error getting file contents: ${error}`
-                        )
-                        return []
-                    }
-                    return {
-                        uri: result.file,
-                        range,
-                        source: 'search',
-                        text,
-                    }
-                }
-            )
-            return (await Promise.all(items)).flat()
-        })
-
-        const allResults = (await Promise.all(r0)).flat()
-
-        if (isAgentTesting) {
-            // Sort results for deterministic ordering for stable tests. Ideally, we
-            // could sort by some numerical score from symf based on how relevant
-            // the matches are for the query.
-            allResults.sort((a, b) => {
-                const byUri = a.uri.fsPath.localeCompare(b.uri.fsPath)
-                if (byUri !== 0) {
-                    return byUri
-                }
-                return a.text.localeCompare(b.text)
-            })
-        }
-
-        return allResults
-    }
-
-    private async searchEmbeddingsLocal(text: string): Promise<ContextItem[]> {
-        if (!this.localEmbeddings) {
-            return []
-        }
-
-        logDebug('SimpleChatPanelProvider', 'getEnhancedContext > searching local embeddings')
-        const contextItems: ContextItem[] = []
-        const embeddingsResults = await this.localEmbeddings.getContext(
-            text,
-            NUM_CODE_RESULTS + NUM_TEXT_RESULTS
-        )
-
-        for (const result of embeddingsResults) {
-            const range = new vscode.Range(
-                new vscode.Position(result.startLine, 0),
-                new vscode.Position(result.endLine, 0)
-            )
-
-            // Filter out ignored files
-            if (!isCodyIgnoredFile(result.uri)) {
-                contextItems.push({
-                    uri: result.uri,
-                    range,
-                    text: result.content,
-                    source: 'embeddings',
-                })
-            }
-        }
-        return contextItems
-    }
-
-    // Note: does not throw error if remote embeddings are not available, just returns empty array
-    private async searchEmbeddingsRemote(text: string): Promise<ContextItem[]> {
-        if (!this.embeddingsClient) {
-            return []
-        }
-        const codebase = await this.codebaseStatusProvider?.currentCodebase()
-        if (!codebase?.remote) {
-            return []
-        }
-        const repoId = await this.embeddingsClient.getRepoIdIfEmbeddingExists(codebase.remote)
-        if (isError(repoId)) {
-            throw new Error(`Error retrieving repo ID: ${repoId}`)
-        }
-        if (!repoId) {
-            return []
-        }
-
-        const workspaceFolder = vscode.workspace.workspaceFolders?.at(0)
-        if (!workspaceFolder) {
-            return []
-        }
-
-        logDebug('SimpleChatPanelProvider', 'getEnhancedContext > searching remote embeddings')
-        const contextItems: ContextItem[] = []
-        const embeddings = await this.embeddingsClient.search(
-            [repoId],
-            text,
-            NUM_CODE_RESULTS,
-            NUM_TEXT_RESULTS
-        )
-        if (isError(embeddings)) {
-            throw new Error(`Error retrieving embeddings: ${embeddings}`)
-        }
-        for (const codeResult of embeddings.codeResults) {
-            const range = new vscode.Range(
-                new vscode.Position(codeResult.startLine, 0),
-                new vscode.Position(codeResult.endLine, 0)
-            )
-            if (!isCodyIgnoredFile(codeResult.uri)) {
-                contextItems.push({
-                    uri: codeResult.uri,
-                    range,
-                    text: codeResult.content,
-                    source: 'embeddings',
-                })
-            }
-        }
-
-        for (const textResult of embeddings.textResults) {
-            const range = new vscode.Range(
-                new vscode.Position(textResult.startLine, 0),
-                new vscode.Position(textResult.endLine, 0)
-            )
-            if (!isCodyIgnoredFile(textResult.uri)) {
-                contextItems.push({
-                    uri: textResult.uri,
-                    range,
-                    text: textResult.content,
-                    source: 'embeddings',
-                })
-            }
-        }
-
-        return contextItems
-    }
-
-    private needsReadmeContext(input: string): boolean {
-        input = input.toLowerCase()
-        const question = extractQuestion(input)
-        if (!question) {
-            return false
-        }
-
-        // split input into words, discarding spaces and punctuation
-        const words = input.split(/\W+/).filter(w => w.length > 0)
-        const bagOfWords = Object.fromEntries(words.map(w => [w, true]))
-
-        const projectSignifiers = [
-            'project',
-            'repository',
-            'repo',
-            'library',
-            'package',
-            'module',
-            'codebase',
-        ]
-        const questionIndicators = ['what', 'how', 'describe', 'explain', '?']
-
-        const workspaceUri = this.editor.getWorkspaceRootUri()
-        if (workspaceUri) {
-            const rootBase = workspaceUri.toString().split('/').at(-1)
-            if (rootBase) {
-                projectSignifiers.push(rootBase.toLowerCase())
-            }
-        }
-
-        let containsProjectSignifier = false
-        for (const p of projectSignifiers) {
-            if (bagOfWords[p]) {
-                containsProjectSignifier = true
-                break
-            }
-        }
-
-        let containsQuestionIndicator = false
-        for (const q of questionIndicators) {
-            if (bagOfWords[q]) {
-                containsQuestionIndicator = true
-                break
-            }
-        }
-
-        return containsQuestionIndicator && containsProjectSignifier
-    }
-
-    private static userAttentionRegexps: RegExp[] = [
-        /editor/,
-        /(open|current|this|entire)\s+file/,
-        /current(ly)?\s+open/,
-        /have\s+open/,
-    ]
-
-    private needsUserAttentionContext(input: string): boolean {
-        const inputLowerCase = input.toLowerCase()
-        // If the input matches any of the `editorRegexps` we assume that we have to include
-        // the editor context (e.g., currently open file) to the overall message context.
-        for (const regexp of ContextProvider.userAttentionRegexps) {
-            if (inputLowerCase.match(regexp)) {
-                return true
-            }
-        }
-        return false
-    }
-
-    private async getReadmeContext(): Promise<ContextItem[]> {
-        // global pattern for readme file
-        const readmeGlobalPattern = '{README,README.,readme.,Readm.}*'
-        const readmeUri = (await vscode.workspace.findFiles(readmeGlobalPattern, undefined, 1)).at(0)
-        if (!readmeUri || isCodyIgnoredFile(readmeUri)) {
-            return []
-        }
-        const readmeDoc = await vscode.workspace.openTextDocument(readmeUri)
-        const readmeText = readmeDoc.getText()
-        const { truncated: truncatedReadmeText, range } = truncateTextNearestLine(
-            readmeText,
-            MAX_BYTES_PER_FILE
-        )
-        if (truncatedReadmeText.length === 0) {
-            return []
-        }
-
-        return [
-            {
-                uri: readmeUri,
-                text: truncatedReadmeText,
-                range: viewRangeToRange(range),
-                source: 'editor',
-            },
-        ]
-    }
+    )
+    return new SimpleChatModel(json.chatModel || modelID, (await Promise.all(messages)).flat(), json.id, json.chatTitle)
 }
 
 export async function contextFilesToContextItems(
@@ -1645,49 +1133,6 @@ export async function contextFilesToContextItems(
     ).filter(isDefined)
 }
 
-function viewRangeToRange(range?: ActiveTextEditorSelectionRange): vscode.Range | undefined {
-    if (!range) {
-        return undefined
-    }
-    return new vscode.Range(range.start.line, range.start.character, range.end.line, range.end.character)
-}
-
-async function newChatModelfromTranscriptJSON(
-    json: TranscriptJSON,
-    modelID: string
-): Promise<SimpleChatModel> {
-    const messages: MessageWithContext[][] = json.interactions.map(
-        (interaction: InteractionJSON): MessageWithContext[] => {
-            return [
-                {
-                    message: {
-                        speaker: 'human',
-                        text: interaction.humanMessage.text,
-                    },
-                    displayText: interaction.humanMessage.displayText,
-                    newContextUsed: deserializedContextFilesToContextItems(
-                        interaction.usedContextFiles,
-                        interaction.fullContext
-                    ),
-                },
-                {
-                    message: {
-                        speaker: 'assistant',
-                        text: interaction.assistantMessage.text,
-                    },
-                    displayText: interaction.assistantMessage.displayText,
-                },
-            ]
-        }
-    )
-    return new SimpleChatModel(
-        json.chatModel || modelID,
-        (await Promise.all(messages)).flat(),
-        json.id,
-        json.chatTitle
-    )
-}
-
 function deserializedContextFilesToContextItems(
     files: ContextFile[],
     contextMessages: ContextMessage[]
@@ -1716,18 +1161,6 @@ function deserializedContextFilesToContextItems(
             source: file.source,
         }
     })
-}
-
-function extractQuestion(input: string): string | undefined {
-    input = input.trim()
-    const q = input.indexOf('?')
-    if (q !== -1) {
-        return input.slice(0, q + 1).trim()
-    }
-    if (input.length < 100) {
-        return input
-    }
-    return undefined
 }
 
 function isAbortError(error: Error): boolean {
@@ -1765,4 +1198,28 @@ function getContextWindowForModel(authStatus: AuthStatus, modelID: string): numb
         return 28000 // 7000 tokens * 4 bytes per token
     }
     return 28000 // assume default to Claude-2-like model
+}
+
+// Select the chat model to use in Chat
+function selectModel(authProvider: AuthProvider, models: ChatModelProvider[]): string {
+    const authStatus = authProvider.getAuthStatus()
+    // Free user can only use the default model
+    if (authStatus.isDotCom && authStatus.userCanUpgrade) {
+        return models[0].model
+    }
+    // Check for the last selected model
+    const lastSelectedModelID = localStorage.get('model')
+    if (lastSelectedModelID) {
+        // If the last selected model exists in the list of models then we return it
+        const model = models.find(m => m.model === lastSelectedModelID)
+        if (model) {
+            return lastSelectedModelID
+        }
+    }
+    // If the user has not selected a model before then we return the default model
+    const defaultModel = models.find(m => m.default) || models[0]
+    if (!defaultModel) {
+        throw new Error('No chat model found in server-provided config')
+    }
+    return defaultModel.model
 }
