+++ vscode/test/integration/chat.test.ts
@@ -26,7 +26,7 @@ suite('Chat', function () {
     test('sends and receives a message', async () => {
         await vscode.commands.executeCommand('cody.chat.panel.new')
         const chatView = await getChatViewProvider()
-        await chatView.handleHumanMessageSubmitted('test', 'hello from the human', 'user', [], false)
+        await chatView.handleNewUserMessage('test', 'hello from the human', 'user', [], false)
 
         assert.match((await getTranscript(0)).displayText || '', /^hello from the human$/)
         await waitUntil(async () =>
@@ -39,7 +39,7 @@ suite('Chat', function () {
         await getTextEditorWithSelection()
         await vscode.commands.executeCommand('cody.chat.panel.new')
         const chatView = await getChatViewProvider()
-        await chatView.handleHumanMessageSubmitted('test', 'hello from the human', 'user', [], false)
+        await chatView.handleNewUserMessage('test', 'hello from the human', 'user', [], false)
 
         // Display text should include file link at the end of message
         assert.match((await getTranscript(0)).displayText || '', /^hello from the human$/)
